Layers

AtomicGraphNets.layers.AGNConv

AtomicGraphNets.layers.AGNPool

AtomicGraphNets.layers.AGNConv — Type

AGNConv(in=>out)

Atomic graph convolutional layer. Almost identical to GCNConv from GeometricFlux but adapted to be most similar to Tian's original AGNN structure, so explicitly has self and convolutional weights separately.

Fields

selfweight::Array{T,2}: weights applied to features at a node

convweight::Array{T,2}: convolutional weights

bias::Array{T,2}: additive bias (second dimension is always 1 because only learnable per-feature, not per-node)

σ::F: activation function (will be applied before reg_norm to outputs), defaults to softplus

Arguments

in::Integer: the dimension of input features.

out::Integer: the dimension of output features.

σ=softplus: activation function

initW=glorot_uniform: initialization function for weights

initb=zeros: initialization function for biases

AtomicGraphNets.layers.AGNPool — Type

Custom pooling layer that outputs a fixed-length feature vector irrespective of input dimensions, for consistent handling of different-sized graphs feeding to fully-connected dense layers afterwards. Adapted from Flux's MeanPool.

It accepts a pooling width and will adjust stride and/or padding such that the output vector length is correct.